Fix vitdodec Linux app compilation
